refactor(dev): move find_repository_root into utils
cli.py and publish.py each carried a byte-identical copy. Keep one in
utils.py and import it in both. publish.py no longer needs `sys`.
